Split action into separate classes
